@@ -67,7 +67,8 @@ AVCaptureDeviceFormat* SelectClosestFormat(AVCaptureDevice* device,
 MacCapturer::MacCapturer(size_t width,
                          size_t height,
                          size_t target_fps,
-                         AVCaptureDevice* device) {
+                         AVCaptureDevice* device)
+    : ScalableVideoTrackSource(ScalableVideoTrackSourceConfig()) {
   RTC_LOG(LS_INFO) << "MacCapturer width=" << width << ", height=" << height
                    << ", target_fps=" << target_fps;
 
