@@ -19,10 +19,12 @@
 #include <api/video/video_rotation.h>
 #include <rtc_base/logging.h>
 
-#include "native_buffer.h"
+// libyuv
+#include <libyuv.h>
 
-ScalableVideoTrackSource::ScalableVideoTrackSource()
-    : AdaptedVideoTrackSource(4) {}
+ScalableVideoTrackSource::ScalableVideoTrackSource(
+    ScalableVideoTrackSourceConfig config)
+    : AdaptedVideoTrackSource(4), config_(config) {}
 ScalableVideoTrackSource::~ScalableVideoTrackSource() {}
 
 bool ScalableVideoTrackSource::is_screencast() const {
@@ -42,11 +44,51 @@ bool ScalableVideoTrackSource::remote() const {
   return false;
 }
 
-void ScalableVideoTrackSource::OnCapturedFrame(
-    const webrtc::VideoFrame& frame) {
+bool ScalableVideoTrackSource::OnCapturedFrame(
+    const webrtc::VideoFrame& video_frame) {
+  webrtc::VideoFrame frame = video_frame;
+
   const int64_t timestamp_us = frame.timestamp_us();
   const int64_t translated_timestamp_us =
       timestamp_aligner_.TranslateTimestamp(timestamp_us, rtc::TimeMicros());
+
+  // 回転が必要
+  if (frame.rotation() != webrtc::kVideoRotation_0) {
+    int width;
+    int height;
+    libyuv::RotationMode mode;
+    switch (frame.rotation()) {
+      case webrtc::kVideoRotation_180:
+        width = frame.width();
+        height = frame.height();
+        mode = libyuv::kRotate180;
+        break;
+      case webrtc::kVideoRotation_90:
+        width = frame.height();
+        height = frame.width();
+        mode = libyuv::kRotate90;
+        break;
+      case webrtc::kVideoRotation_270:
+      default:
+        width = frame.height();
+        height = frame.width();
+        mode = libyuv::kRotate270;
+        break;
+    }
+
+    rtc::scoped_refptr<webrtc::I420Buffer> rotated =
+        webrtc::I420Buffer::Create(width, height);
+    rtc::scoped_refptr<webrtc::I420BufferInterface> src =
+        frame.video_frame_buffer()->ToI420();
+    libyuv::I420Rotate(src->DataY(), src->StrideY(), src->DataU(),
+                       src->StrideU(), src->DataV(), src->StrideV(),
+                       rotated->MutableDataY(), rotated->StrideY(),
+                       rotated->MutableDataU(), rotated->StrideU(),
+                       rotated->MutableDataV(), rotated->StrideV(),
+                       frame.width(), frame.height(), mode);
+    frame.set_video_frame_buffer(rotated);
+    frame.set_rotation(webrtc::kVideoRotation_0);
+  }
 
   int adapted_width;
   int adapted_height;
@@ -57,16 +99,17 @@ void ScalableVideoTrackSource::OnCapturedFrame(
   if (!AdaptFrame(frame.width(), frame.height(), timestamp_us, &adapted_width,
                   &adapted_height, &crop_width, &crop_height, &crop_x,
                   &crop_y)) {
-    return;
+    return false;
   }
 
-  if (UseNativeBuffer() && frame.video_frame_buffer()->type() ==
-                               webrtc::VideoFrameBuffer::Type::kNative) {
-    NativeBuffer* frame_buffer =
-        dynamic_cast<NativeBuffer*>(frame.video_frame_buffer().get());
-    frame_buffer->SetScaledSize(adapted_width, adapted_height);
+  if (config_.on_frame) {
+    config_.on_frame(frame);
+  }
+
+  if (frame.video_frame_buffer()->type() ==
+      webrtc::VideoFrameBuffer::Type::kNative) {
     OnFrame(frame);
-    return;
+    return true;
   }
 
   rtc::scoped_refptr<webrtc::VideoFrameBuffer> buffer =
@@ -86,4 +129,6 @@ void ScalableVideoTrackSource::OnCapturedFrame(
               .set_rotation(frame.rotation())
               .set_timestamp_us(translated_timestamp_us)
               .build());
+
+  return true;
 }
