@@ -17,6 +17,7 @@
 #endif
 
 #if USE_JETSON_ENCODER
+#include "hwenc_jetson/jetson_video_decoder.h"
 #include "hwenc_jetson/jetson_video_encoder.h"
 #endif
 
@@ -239,10 +240,18 @@ struct VideoCodecInfo {
     if (JetsonVideoEncoder::IsSupportedVP8()) {
       info.vp8_encoders.push_back(Type::Jetson);
     }
-    info.vp8_decoders.push_back(Type::Jetson);
-    info.vp9_decoders.push_back(Type::Jetson);
+    if (JetsonVideoDecoder::IsSupportedVP8()) {
+      info.vp8_decoders.push_back(Type::Jetson);
+    }
     if (JetsonVideoEncoder::IsSupportedVP9()) {
       info.vp9_encoders.push_back(Type::Jetson);
+    }
+    info.vp9_decoders.push_back(Type::Jetson);
+    if (JetsonVideoEncoder::IsSupportedAV1()) {
+      info.av1_encoders.push_back(Type::Jetson);
+    }
+    if (JetsonVideoDecoder::IsSupportedAV1()) {
+      info.av1_decoders.push_back(Type::Jetson);
     }
 #endif
 
